Move wrapper types into separate module
