+++ src/lib.rs
@@ -133,14 +133,18 @@
 //! # let mut sht = shtcx::shtc3(I2cdev::new("/dev/i2c-1").unwrap(), Delay);
 //! sht.reset().unwrap();
 //! ```
-
 #![deny(unsafe_code, missing_docs)]
 #![cfg_attr(not(test), no_std)]
+
+mod types;
 
 use core::marker::PhantomData;
 
 use embedded_hal::blocking::delay::{DelayMs, DelayUs};
 use embedded_hal::blocking::i2c::{Read, Write, WriteRead};
+
+pub use types::*;
+
 
 /// Whether temperature or humidity is returned first when doing a measurement.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
@@ -254,57 +258,6 @@ pub struct ShtCx<S: ShtSensor, I2C, D> {
     delay: D,
     /// The I²C device address.
     address: u8,
-}
-
-/// A temperature measurement.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Temperature(i32);
-
-/// A humidity measurement.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Humidity(i32);
-
-/// A combined temperature / humidity measurement.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Measurement {
-    /// The measured temperature.
-    pub temperature: Temperature,
-    /// The measured humidity.
-    pub humidity: Humidity,
-}
-
-impl Temperature {
-    /// Create a new `Temperature` from a raw measurement result.
-    fn from_raw(raw: u16) -> Self {
-        Self(convert_temperature(raw))
-    }
-
-    /// Return temperature in milli-degrees celsius.
-    pub fn as_millidegrees_celsius(&self) -> i32 {
-        self.0
-    }
-
-    /// Return temperature in degrees celsius.
-    pub fn as_degrees_celsius(&self) -> f32 {
-        self.0 as f32 / 1000.0
-    }
-}
-
-impl Humidity {
-    /// Create a new `Humidity` from a raw measurement result.
-    fn from_raw(raw: u16) -> Self {
-        Self(convert_humidity(raw))
-    }
-
-    /// Return relative humidity in 1/1000 %RH.
-    pub fn as_millipercent(&self) -> i32 {
-        self.0
-    }
-
-    /// Return relative humidity in %RH.
-    pub fn as_percent(&self) -> f32 {
-        self.0 as f32 / 1000.0
-    }
 }
 
 /// Create a new instance of the driver for the SHTC1.
@@ -544,24 +497,6 @@ fn crc8(data: &[u8]) -> u8 {
         }
     }
     crc
-}
-
-/// Convert raw temperature measurement to milli-degrees celsius.
-///
-/// Formula (datasheet 5.11): -45 + 175 * (val / 2^16),
-/// optimized for fixed point math.
-#[inline]
-fn convert_temperature(temp_raw: u16) -> i32 {
-    (((temp_raw as u32) * 21875) >> 13) as i32 - 45000
-}
-
-/// Convert raw humidity measurement to relative humidity.
-///
-/// Formula (datasheet 5.11): 100 * (val / 2^16),
-/// optimized for fixed point math.
-#[inline]
-fn convert_humidity(humi_raw: u16) -> i32 {
-    (((humi_raw as u32) * 12500) >> 13) as i32
 }
 
 #[cfg(test)]
@@ -835,16 +770,6 @@ mod tests {
             assert_eq!(err, Error::I2c(MockError::Io(ErrorKind::Other)));
             sht.destroy().done();
         }
-
-        /// Test conversion of raw measurement results into °C and %RH.
-        #[test]
-        fn measurement_conversion() {
-            // Datasheet setion 5.11 "Conversion of Sensor Output"
-            let temperature = convert_temperature(((0b0110_0100 as u16) << 8) | 0b1000_1011);
-            let humidity = convert_humidity(((0b1010_0001 as u16) << 8) | 0b0011_0011);
-            assert_eq!(temperature, 23730);
-            assert_eq!(humidity, 62968);
-        }
     }
 
     mod power_management {
@@ -878,36 +803,6 @@ mod tests {
             let mut sht = shtc3(mock, NoopDelay);
             sht.reset().unwrap();
             sht.destroy().done();
-        }
-    }
-
-    mod wrapper_types {
-        use super::*;
-
-        /// Test conversion of raw measurement results into °C.
-        #[test]
-        fn test_convert_temperature() {
-            let test_data = [
-                (0x0000, -45000),
-                // Datasheet setion 5.11 "Conversion of Sensor Output"
-                (((0b0110_0100 as u16) << 8) | 0b1000_1011, 23730),
-            ];
-            for td in &test_data {
-                assert_eq!(convert_temperature(td.0), td.1);
-            }
-        }
-
-        /// Test conversion of raw measurement results into %RH.
-        #[test]
-        fn test_convert_humidity() {
-            let test_data = [
-                (0x0000, 0),
-                // Datasheet setion 5.11 "Conversion of Sensor Output"
-                (((0b1010_0001 as u16) << 8) | 0b0011_0011, 62968),
-            ];
-            for td in &test_data {
-                assert_eq!(convert_humidity(td.0), td.1);
-            }
         }
     }
 }
